deep_learning/flock_analyzer/requirements.txt
torch
opencv-python
numpy
scikit-learn
ultralytics
Pillow